// minhadocao-app/src/types/CachorroResponseDTO.ts
export interface CachorroResponseDTO {
    id: string;
    nome: string;
    categoria: string;
    idade: string;
    dataNascimento: string;
    status: string;
    descricao: string;
    urlImagem: string;
    adotante: string;
    docAdotante: string;
    emailAdotante: string;
    raca: string;
  }